updated task 1 and task 2 pages with qr fn
