+++ client/src/components/Home/Home.js
@@ -1,11 +1,6 @@
 import './Home.css';
 import React, {useEffect, useState} from "react";
 import axios from "axios";
-import VideoReader from "../VideoReader/VideoReader";
-import { styled } from '@mui/material/styles';
-import Paper from '@mui/material/Paper';
-import Grid from '@mui/material/Unstable_Grid2';
-import Box from '@mui/material/Box';
 
 function Home() {
     const [test, setTest] = useState("");
@@ -18,35 +13,16 @@ function Home() {
         })
     }, []);
 
-    const Item = styled(Paper)(({ theme }) => ({
-        backgroundColor: theme.palette.mode === 'dark' ? '#1A2027' : '#fff',
-        ...theme.typography.body2,
-        padding: theme.spacing(1),
-        textAlign: 'center',
-        color: theme.palette.text.secondary,
-        minHeight: '40vh'
-      }));
-
     return (
-        <Box sx={{ flexGrow: 1 }} className="Home">
-            <Grid container spacing={2}>
-                <Grid xs={6}>
-                    <Item>
-                        {/* Component QR Reader */}
-                        <VideoReader/>
-                    </Item>
-                </Grid>
-                <Grid xs={6}>
-                    <Item>Component Here</Item>
-                </Grid>
-                <Grid xs={6}>
-                    <Item>Component Here</Item>
-                </Grid>
-                <Grid xs={6}>
-                    <Item>Component Here</Item>
-                </Grid>
-            </Grid>
-        </Box>
+        <div className="Home">
+            <header className='app-header'>
+                TMAV
+            </header>
+            <br/>
+            <p>
+               Test Data: {test}
+            </p>
+        </div>
     );
 }
