+++ client/src/components/Task2/Task2.css
@@ -1,0 +1,7 @@
+.Task2 {
+    background-color: #282c34;
+    min-height: 100vh;
+    font-size: calc(10px + 2vmin);
+    color: white;
+    padding: 10px
+  }
