+++ server/app.py
@@ -32,7 +32,7 @@ def validate_qr():
     # Checks format of raw_qr_string if it conforms to the type
     # Saves QR formatted class to variable
     # Returns success if valid
-    jsonResponse = request.get_json() # paarsable dictionary
+    jsonResponse = request.get_json() # parsable dictionary
     print(jsonResponse)
     return {'status': 'SUCCESS'}
